+++ apps/web/app/future/auth/setup/page.tsx
@@ -1,12 +1,13 @@
-import Setup from "@pages/auth/setup";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { WithLayout } from "app/layoutHOC";
-import type { InferGetServerSidePropsType } from "next";
 
 import { getServerSideProps } from "@server/lib/setup/getServerSideProps";
+
+import Setup from "~/auth/setup-view";
+import type { PageProps } from "~/auth/setup-view";
 
 export default WithLayout({
   getLayout: null,
   Page: Setup,
-  getData: withAppDirSsr<InferGetServerSidePropsType<typeof getServerSideProps>>(getServerSideProps),
+  getData: withAppDirSsr<PageProps>(getServerSideProps),
 })<"P">;
