+++ apps/web/app/future/auth/signin/page.tsx
@@ -1,13 +1,14 @@
-import signin from "@pages/auth/signin";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { WithLayout } from "app/layoutHOC";
-import type { InferGetServerSidePropsType } from "next";
 
 import { getServerSideProps } from "@server/lib/auth/signin/getServerSideProps";
 
+import SignIn from "~/auth/signin-view";
+import type { PageProps } from "~/auth/signin-view";
+
 export default WithLayout({
   getLayout: null,
-  Page: signin,
+  Page: SignIn,
   // @ts-expect-error TODO: fix this
-  getData: withAppDirSsr<InferGetServerSidePropsType<typeof getServerSideProps>>(getServerSideProps),
+  getData: withAppDirSsr<PageProps>(getServerSideProps),
 })<"P">;
