+++ apps/web/app/future/auth/sso/direct/page.tsx
@@ -1,11 +1,13 @@
-import DirectSSOLogin from "@pages/auth/sso/direct";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { WithLayout } from "app/layoutHOC";
 
 import { getServerSideProps } from "@server/lib/auth/sso/direct/getServerSideProps";
 
+import type { SSODirectPageProps } from "~/auth/sso/direct-view";
+import SSODirectView from "~/auth/sso/direct-view";
+
 export default WithLayout({
   getLayout: null,
-  Page: DirectSSOLogin,
-  getData: withAppDirSsr(getServerSideProps),
+  Page: SSODirectView,
+  getData: withAppDirSsr<SSODirectPageProps>(getServerSideProps),
 })<"P">;
