+++ apps/web/app/future/insights/page.tsx
@@ -1,10 +1,11 @@
-import LegacyPage from "@pages/insights/index";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
 
 import { getServerSideProps } from "@lib/insights/getServerSideProps";
-import { type inferSSRProps } from "@lib/types/inferSSRProps";
+
+import type { PageProps } from "~/insights/insights-view";
+import InsightsPage from "~/insights/insights-view";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -12,6 +13,6 @@ export const generateMetadata = async () =>
     (t) => t("insights_subtitle")
   );
 
-const getData = withAppDirSsr<inferSSRProps<typeof getServerSideProps>>(getServerSideProps);
+const getData = withAppDirSsr<PageProps>(getServerSideProps);
 
-export default WithLayout({ getLayout: null, getData, Page: LegacyPage });
+export default WithLayout({ getLayout: null, getData, Page: InsightsPage });
