+++ apps/web/app/future/org/[orgSlug]/page.tsx
@@ -1,5 +1,5 @@
 import { withAppDirSsr } from "app/WithAppDirSsr";
-import type { Params, SearchParams } from "app/_types";
+import type { PageProps as _PageProps } from "app/_types";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
 import { cookies, headers } from "next/headers";
@@ -10,13 +10,7 @@ import { getServerSideProps } from "@lib/team/[slug]/getServerSideProps";
 import type { PageProps } from "~/team/team-view";
 import TeamPage from "~/team/team-view";
 
-export const generateMetadata = async ({
-  params,
-  searchParams,
-}: {
-  params: Params;
-  searchParams: SearchParams;
-}) => {
+export const generateMetadata = async ({ params, searchParams }: _PageProps) => {
   const props = await getData(buildLegacyCtx(headers(), cookies(), params, searchParams));
   const teamName = props.team.name || "Nameless Team";
