+++ apps/web/app/future/reschedule/[uid]/page.tsx
@@ -1,10 +1,10 @@
-import { getServerSideProps } from "@pages/reschedule/[uid]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import type { PageProps } from "app/_types";
 import { _generateMetadata } from "app/_utils";
 import { headers, cookies } from "next/headers";
 
 import { buildLegacyCtx } from "@lib/buildLegacyCtx";
+import { getServerSideProps } from "@lib/reschedule/[uid]/getServerSideProps";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
