+++ apps/web/app/layoutHOC.tsx
@@ -51,6 +51,7 @@ export function WithLayout<T extends Record<string, any>>({
         requiresLicense={requiresLicense || !!(Page && "requiresLicense" in Page && Page.requiresLicense)}
         nonce={nonce}
         themeBasis={null}
+        isThemeSupported={!!(Page && "isThemeSupported" in Page && Page.isThemeSupported)}
         isBookingPage={isBookingPage || !!(Page && "isBookingPage" in Page && Page.isBookingPage)}
         {...props}>
         {pageWithServerLayout}
