+++ apps/web/pages/reschedule/[uid]/embed.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import withEmbedSsr from "@lib/withEmbedSsr";
 
 import { getServerSideProps as _getServerSideProps } from "../[uid]";
